Basic Start Menu
A basic start menu, with buttons for playing the first level and
quitting. IT could look nicer and it doesn't have a logo or the name of
the game anywhere. Also could add loading a game, or jumping to a
specific level depending on how we want to design the game. The tutorial
could also be a level itself.