Non-Functional Inventory
It can slide up and down, by pressing on the little button that says
inventory. Probably going to go back and tweak the looks a bit it's
kinda ugly right now. All images were made with Gimp, saved in a png and
the default gimp save. Also moved all the builds to a folder.