Basically Working Inventory
The Inventory Script can move items around. To create a new item, create
a png file, and save it to Images, in the proper equipment folder. Then,
go into the EquipCategories script, and edit the appropriate array,
adding a float value based on how good the weapon is. The array is in
the same order as the alphabetical order of the images in each folder,
and should be organized from least to greatest.

TODO: Make it so equipped weapons do stuff, and change it so only
objects of a certain type can be equipped.